Npmignore, submitoutput rework
