+++ .npmignore
@@ -1,0 +1,21 @@
+themeprovider-example
+node_modules
+public
+DevProfile
+.vscode
+**/cypress/videos
+**/cypress/screenshots
+**/constants/*JsSource.ts
+.cache
+.github
+debug.log
+now.json
+.prettierignore
+.prettierrc.yaml
+static
+bin
+example
+src
+cypress.json
+index.tsx
+tsconfig.json
