+++ src/components/SubmissionOutput.tsx
@@ -1,6 +1,6 @@
-import { CircularProgress } from "@material-ui/core"
 import React, { useState } from "react"
 import { useTranslation } from "react-i18next"
+import useStyles from "../hooks/useStyles"
 
 import { TestResultObject } from "../types"
 import Help from "./Help"
@@ -9,7 +9,6 @@ import {
   OutputBox,
   OutputButton,
   OutputColor,
-  OutputHeaderText,
   OutputFooterWithPercentage,
   OutputHeader,
 } from "./OutputBox"
@@ -21,6 +20,7 @@ interface SubmissionOutputProps {
   getPasteLink: () => Promise<string>
   pasteDisabled: boolean
   onClose: () => void
+  onSubmit: () => void
   outputHeight?: string
   testResults: TestResultObject
 }
@@ -31,12 +31,14 @@ const SubmissionOutput: React.FunctionComponent<SubmissionOutputProps> = (
   const {
     getPasteLink,
     onClose,
+    onSubmit,
     testResults,
     pasteDisabled,
     outputHeight,
   } = props
   const [t] = useTranslation()
   const [showAllTests, setShowAllTests] = useState(false)
+  const classes = useStyles()
 
   const percentage = Math.round(
     (100 * testResults.testCases.filter((x) => x.passed).length) /
@@ -72,7 +74,17 @@ const SubmissionOutput: React.FunctionComponent<SubmissionOutputProps> = (
           testResults.testCases.length === 1 &&
           (!testResults.allTestsPassed ?? false)
         }
-      />
+      >
+        {!testResults.allTestsPassed && (
+          <OutputButton
+            disabled={false}
+            label={t("button.submit")}
+            onClick={onSubmit}
+            dataCy="submit-btn"
+            className={classes.blueButton}
+          />
+        )}
+      </OutputFooterWithPercentage>
     </OutputBox>
   )
 }
